chore(refactor): move bins to dir & text maze impl to lib
