+++ src/traits/mod.rs
@@ -1,0 +1,5 @@
+mod maze;
+mod robot;
+
+pub use maze::{Maze, MazeError};
+pub use robot::{Robot, RobotInternal};
